// src/types/crypto.ts
export interface Crypto {
  id: number;
  createdat: string;
  updatedat: string;
  name: string;
  symbol: string;
  decimals: number;
  address: string;
  writer: any;
  active: number;
  nettype: string;
  istoken: number;
  urllogo: string;
  isdefault: any;
  uuid: string;
  deployer: any;
  isadminadded: any;
  totalsupply: any;
  typestr: string;
  nation: any;
  typecf: string;
  policy: any;
  mindeposit: string;
  maxdeposit: any;
  minwithdraw: string;
}